Reorganized modules/contentItems/actions folder to match the other modules.
