+++ app/modules/contentItems/actions/apiSaga/index.js
@@ -1,0 +1,6 @@
+// @flow
+
+const apiSagaActions = {
+};
+
+export default apiSagaActions;
