+++ app/modules/contentItems/actions/reducer/removeFromState.js
@@ -1,6 +1,6 @@
 // @flow
 
-import * as a from '../actionTypes';
+import * as a from '../../actionTypes';
 
 const removeFromState = (
   id: string,
