+++ app/modules/contentItems/actions/reducer/setMultipleInState.js
@@ -1,7 +1,7 @@
 // @flow
 
-import * as a from '../actionTypes';
-import * as m from '../model';
+import * as a from '../../actionTypes';
+import * as m from '../../model';
 
 const setMultipleInState = (
   contentItems: $ReadOnlyArray<m.ContentItem>,
