+++ app/modules/contentItems/actions/reducer/setMultipleInState.test.js
@@ -2,8 +2,8 @@
 
 import { dummyContentItemData as dummyData } from 'lib/testResources';
 
-import * as a from '../actionTypes';
-import * as m from '../model';
+import * as a from '../../actionTypes';
+import * as m from '../../model';
 
 import actions from '.';
