+++ app/modules/contentItems/actions/taskSaga/add.js
@@ -1,7 +1,7 @@
 // @flow
 
-import * as a from '../actionTypes';
-import * as m from '../model';
+import * as a from '../../actionTypes';
+import * as m from '../../model';
 
 const add = (
   type: m.ContentItemType,
