+++ app/modules/contentItems/actions/taskSaga/edit.test.js
@@ -1,7 +1,7 @@
 // @flow
 
-import * as a from '../actionTypes';
-import * as m from '../model';
+import * as a from '../../actionTypes';
+import * as m from '../../model';
 
 import actions from '.';
