+++ app/modules/contentItems/actions/taskSaga/indent.js
@@ -1,6 +1,6 @@
 // @flow
 
-import * as a from '../actionTypes';
+import * as a from '../../actionTypes';
 
 const indent = (id: string): a.IndentAction => {
   return {
