+++ app/modules/contentItems/actions/taskSaga/remove.js
@@ -1,6 +1,6 @@
 // @flow
 
-import * as a from '../actionTypes';
+import * as a from '../../actionTypes';
 
 const remove = (id: string): a.RemoveAction => {
   return {
