+++ app/modules/contentItems/actions/taskSaga/remove.test.js
@@ -1,6 +1,6 @@
 // @flow
 
-import * as a from '../actionTypes';
+import * as a from '../../actionTypes';
 
 import actions from '.';
