+++ app/modules/contentItems/actions/taskSaga/toggleEditing.js
@@ -1,6 +1,6 @@
 // @flow
 
-import * as a from '../actionTypes';
+import * as a from '../../actionTypes';
 
 const toggleEditing = (
   id: string,
